requirements.txt
Flask
requests==2.32.3
gunicorn
python-dotenv==1.1.0
waitress

